refactor(app): LabwareStackModal to molecules
