+++ app/src/local-resources/modules/index.ts
@@ -1,1 +1,2 @@
 export * from './getModulePrepCommands'
+export * from './getModuleImage'
