+++ app/src/organisms/Devices/ProtocolRun/SetupModuleAndDeck/SetupModulesList.tsx
@@ -43,7 +43,10 @@ import {
 } from '/app/resources/runs'
 import { ModuleSetupModal } from '../../../ModuleCard/ModuleSetupModal'
 import { ModuleWizardFlows } from '../../../ModuleWizardFlows'
-import { getModulePrepCommands } from '/app/local-resources/modules'
+import {
+  getModulePrepCommands,
+  getModuleImage,
+} from '/app/local-resources/modules'
 import { getModuleTooHot } from '/app/transformations/modules'
 import {
   useModuleRenderInfoForProtocolById,
@@ -52,7 +55,6 @@ import {
 import { LocationConflictModal } from '/app/organisms/LocationConflictModal'
 import { OT2MultipleModulesHelp } from './OT2MultipleModulesHelp'
 import { UnMatchedModuleWarning } from './UnMatchedModuleWarning'
-import { getModuleImage } from './utils'
 
 import type {
   CutoutConfig,
